requirements.txt
PyYAML         # translate YAML to python Dict
pymongo        # connect to mongodb
python-dotenv  # load dotenv files
